SEFINI AL 90%

+++ OSDEPYM/src/main/java/com/osdepym/hibernate/dao/Dao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface Dao<T>{
 	
-	T get(long id);
+	T get(Integer id);
     
     List<T> getAll();
